Implemented markImage()
This js function is called when the user selects one of the photos on the web page. This will mark the photo as selected and also place a green border around the image to notate the selection to the user. The marked images have an alt value of "1" which will be used to find them to submit their ID's to the api.