(uploader.go) Use channels instead of waitgroup
Uploader.go now uses waitgroups to keep the max number of concurrent goroutines. Before, the code would hit the max number of routines then wait until they all finished. Now the goroutines will send a message in the channel that will signal its completion. The loop that is sitting there waiting will recieve that signal and it will then start another goroutine so to always keep the max number running to use the CPU to it's highest potential.