Images now display in order every time
The issue was that the thumbnail key and url were being stored in a map which is an unordered data type. When the code later iterated through the map, it would be in a different order at every runtime. The map has now been replaced with a struct "Thumbnail". This struct has a property for the key and the url. Now when the html is created, it will be in the same order every time.