Gallery pagination. Broke js and css. Needs fix
Now the galleries have a /<number> at the end representing the page. Page 0 is the first page, so on and so forth. As of now the js and css are broken due to the changes but I currently can't seem to figure out how to get the static file mappings with Gin to play nicely.