Fixed pagination, cookie timeout. Added page num, arrow nav
Fixed the issue with javascript and css having the page html appended to them. Fixed issue where the browser cookie would delete itself after 15 minutes regardless of activity. Added arrow buttons to the shoot interface that allow the user to navigate between different pages. Added a part to the navbar that tells the user what page they are on.