Refactored variables and handled errors
There were a lot of inconsistencies such as variables that were written in snake case rather than camel case like the rest of the project. Camel case is also the standard case for Go projects. There were also several errors that were not handled in any way. They are discarded but now the errors are at least acknowledged.