Replaced submit with save. Added save status indicator
Instead of "submitting" your selections, you now "save" them. The backend is still the exact same, it just looks different visibly on the site. The purpose behind this is to prep for the addition of cookies that will track selections between pages and also keep selections between logins and etc.