Fixed some front end timing async issues.
Fixed async issues with front end. Also removed unused function "createSecureCookie()". Issues may still arise when switching between ports as port 443 seems to add a "." to the front of the domain name when used. When port 8080 is used, the dot does not seem to be there. Needs to be diagnosed.