Home page
Now when you log in, you are presented with a page that shows all of the shoots you have on your user. You can then click on the shoot and it will take you to the page. There is also now a home button that will take you back to the home page.