Fetch user route in debug mode
When in debug mode, a route is available that allows you to get the full record of a user from the database. This is restricted to debug mode as there is no auth or anything like that and doesn't really have a purpose outside of just debugging things during development.