Thumbnails can be cross account
Previously, the entry in the database for the thumbnail was only the file name and the rest was computed in the backend code. The way it was done meant that the shoot thumbnail would only be correctly generated if that specific user was doing it. I was creating a guest account and found this to be a problem. So now the thumbnail gets an actual full S3 bucket path to the file.