Fix static file browser caching
The previous commit erroneously set the header for browser caching on every request that came to the server. While application performance increase considerably, it led to various unexpected behaviors, most particularly with the cookies. This commit moves the header set in the middleware to be only if the middleware is actually sending a cached static file.